+++ src/fundamental.c
@@ -1,0 +1,9 @@
+#include "tega/fundamental.h"
+
+TFUN_Bytes TFUN_Bytes_new() {
+    TFUN_Bytes bytes;
+    bytes.ptr = NULL;
+    bytes.len = 0;
+    bytes.cap = 0;
+    return bytes;
+}