+++ tests/fundamental/test_tfun_bytes.c
@@ -1,0 +1,12 @@
+#include <tega/fundamental.h>
+
+#include "../tests.h"
+#include "test_fundamental.h"
+
+static void bytesNew() {
+    TEST("TFUN_Bytes_new");
+}
+
+void testTFUN_Bytes() {
+    bytesNew();
+}