+++ tests/test_fundamental.c
@@ -1,0 +1,6 @@
+#include "fundamental/test_fundamental.h"
+#include "tests.h"
+
+void testFundamental() {
+    testTFUN_Bytes();
+}